requirements.txt
flask>=3.0
requests>=2.31
python-dotenv>=1.0
groq>=0.9
gunicorn>=21.0
